% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.C.collapsed.r
\name{augment.C.collapsed}
\alias{augment.C.collapsed}
\title{augment.C.collapsed}
\usage{
augment.C.collapsed(w_sums, Vobs, kappa, theta1, r, g.fixed)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
augment.C.collapsed
}
