% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.C.r
\name{augment.C}
\alias{augment.C}
\title{augment.C}
\usage{
augment.C(pobs, Vobs, X, kappa, theta1, r, g.fixed)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
augment.C
}
