% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.X.r
\name{augment.X}
\alias{augment.X}
\title{augment.X}
\usage{
augment.X(pobs_vec, Vobs, Vobs_L, Vobs_R, cur.par.X, dist.X, C, collapsed.g)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
augment.X
}
