% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.X_rcpp.r
\name{augment.X_rcpp}
\alias{augment.X_rcpp}
\title{augment.X_rcpp}
\usage{
augment.X_rcpp(
  pobs_vec,
  Vobs,
  Vobs_L,
  Vobs_R,
  cur.par.X,
  dist.X,
  C,
  collapsed.g
)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
augment.X_rcpp
}
