% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayespim_package.R
\docType{package}
\name{BayesPIM-package}
\alias{BayesPIM}
\alias{BayesPIM-package}
\title{BayesPIM: Bayesian Prevalence-Incidence Mixture Model}
\description{
Models time-to-event data from interval-censored screening studies. It accounts for latent prevalence at baseline and incorporates misclassification due to imperfect test sensitivity. For usage details, see the package vignette "BayesPIM_intro". Further details can be found in Klausch, Lissenberg-Witte and Coupé (2026) \doi{10.1002/sim.70433}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasklausch2/BayesPIM}
  \item Report bugs at \url{https://github.com/thomasklausch2/BayesPIM/issues}
}

}
\author{
\strong{Maintainer}: Thomas Klausch \email{t.klausch@amsterdamumc.nl}

Authors:
\itemize{
  \item Thomas Klausch \email{t.klausch@amsterdamumc.nl}
}

}
\keyword{internal}
