% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.mcmclists.r
\name{bind.mcmclists}
\alias{bind.mcmclists}
\title{bind.mcmclists}
\usage{
bind.mcmclists(pr, r)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
bind.mcmclists
}
