% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_beta.r
\name{fc_beta}
\alias{fc_beta}
\title{fc_beta}
\usage{
fc_beta(X, y, sig_inv_Xt, sig_inv)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
fc_beta
}
