% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_w_par.exp_Haar.r
\name{fc_w_par.exp_Haar}
\alias{fc_w_par.exp_Haar}
\title{fc_w_par.exp_Haar}
\usage{
fc_w_par.exp_Haar(y, X, sig_inv_Xt)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
fc_w_par.exp_Haar
}
