% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lobs_2S.r
\name{Lobs_2S}
\alias{Lobs_2S}
\title{Lobs_2S}
\usage{
Lobs_2S(est, mod, log.scale = T, sumup = T)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
Lobs_2S
}
