% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_gengamma.r
\name{log_likelihood_gengamma}
\alias{log_likelihood_gengamma}
\title{log_likelihood_gengamma}
\usage{
log_likelihood_gengamma(params, x, z)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
log_likelihood_gengamma
}
