% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.up.mat.r
\name{look.up.mat}
\alias{look.up.mat}
\title{look.up.mat}
\usage{
look.up.mat(L, a)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
look.up.mat
}
