% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhstep.aft.r
\name{mhstep.aft}
\alias{mhstep.aft}
\title{mhstep.aft}
\usage{
mhstep.aft(
  x,
  t,
  Z,
  tau,
  sig.prior,
  k.prior,
  prop.var = diag(1, length(x)),
  dist,
  fix.sigma = FALSE,
  fix.k = F,
  beta.prior = "t"
)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
mhstep.aft
}
