% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst.aft.r
\name{pst.aft}
\alias{pst.aft}
\title{pst.aft}
\usage{
pst.aft(par, t, Z, tau, sig.prior, k.prior, dist, beta.prior = "t")
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
pst.aft
}
