% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst.kappa.noprev.r
\name{pst.kappa.noprev}
\alias{pst.kappa.noprev}
\title{pst.kappa.noprev}
\usage{
pst.kappa.noprev(Vobs, j_, a = 1, b = 1)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
pst.kappa.noprev
}
