% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst.kappa.r
\name{pst.kappa}
\alias{pst.kappa}
\title{pst.kappa}
\usage{
pst.kappa(Vobs, j_, a = 1, b = 1, g, r, g.fixed)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
pst.kappa
}
