% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pst.X.2S.r
\name{pst.X.2S}
\alias{pst.X.2S}
\title{pst.X.2S}
\usage{
pst.X.2S(par, dist, S, d, L, R)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
pst.X.2S
}
