% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.trdist.r
\name{r.trdist}
\alias{r.trdist}
\title{r.trdist}
\usage{
r.trdist(par, a = 0, b = Inf, dist, tol = 1e-08)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
r.trdist
}
