% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtruncnorm_inv.r
\name{rtruncnorm_inv}
\alias{rtruncnorm_inv}
\title{rtruncnorm_inv}
\usage{
rtruncnorm_inv(n, mu = 0, sd = 1, a = -Inf, b = Inf)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
rtruncnorm_inv
}
