% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.ppd.vanilla.r
\name{sample.ppd.vanilla}
\alias{sample.ppd.vanilla}
\title{sample.ppd.vanilla}
\usage{
sample.ppd.vanilla(par.list, Z.X, dist.X, s)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
sample.ppd.vanilla
}
