% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.ppd.xstar.r
\name{sample.ppd.xstar}
\alias{sample.ppd.xstar}
\title{sample.ppd.xstar}
\usage{
sample.ppd.xstar(par.list, Z.X, Z.W, dist.X, s, g.fixed, type)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
sample.ppd.xstar
}
