% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.par.gengamma.r
\name{trans.par.gengamma}
\alias{trans.par.gengamma}
\title{trans.par.gengamma}
\usage{
trans.par.gengamma(Z1, par)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
trans.par.gengamma
}
