% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.par.ind.norm.r
\name{trans.par.ind.norm}
\alias{trans.par.ind.norm}
\title{trans.par.ind.norm}
\usage{
trans.par.ind.norm(p, v, Z1)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
trans.par.ind.norm
}
