% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.par.norm.r
\name{trans.par.norm}
\alias{trans.par.norm}
\title{trans.par.norm}
\usage{
trans.par.norm(p1, p2, v1, v2, cr, Z1, Z2, a)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
trans.par.norm
}
