% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.mcmc.r
\name{trim.mcmc}
\alias{trim.mcmc}
\title{trim.mcmc}
\usage{
trim.mcmc(obj, burnin = 1, end = nrow(as.matrix(obj[1])), thining = 1)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
trim.mcmc
}
