% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_to_vobs.r
\name{v_to_vobs}
\alias{v_to_vobs}
\title{v_to_vobs}
\usage{
v_to_vobs(V, X, kappa, C, baseline.test)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
v_to_vobs
}
