% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes2S_v6_seq.r
\name{bayes.2S_seq}
\alias{bayes.2S_seq}
\title{bayes.2S_seq}
\usage{
bayes.2S_seq(
  Vobs,
  kappa,
  Z.X = NULL,
  Z.W = NULL,
  r = rep(0, length(Vobs)),
  parallel = T,
  update.kappa = F,
  prev = T,
  ndraws = 1000,
  chains = 3,
  thining = 1,
  prop.sd.X = NULL,
  naive.run.prop.sd.X = prop.sd.X,
  beta.prior.X = 4,
  sig.prior.X = 10,
  tau.w = 1,
  k.prior = sqrt(10),
  fix.sigma.X = F,
  prev.run = NULL,
  dist.X = "weibull",
  update.burnin = T,
  beta.prior = "t",
  vanilla = F,
  ndraws.naive = 5000,
  update.till.converge = F,
  ndraws.update = NULL,
  maxit = Inf,
  conv.crit = "upper",
  kappa.prior = NULL,
  fix.k = F,
  par.exp = T,
  collapsed.g = F
)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
bayes.2S_seq
}
