% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPIM_package.r
\docType{package}
\name{BayesPIM-package}
\alias{BayesPIM}
\alias{BayesPIM-package}
\title{BayesPIM: Bayesian Prevalence-Incidence Mixture Model}
\description{
Models time-to-event data from interval-censored screening studies. It accounts for latent prevalence at baseline and incorporates misclassification due to imperfect test sensitivity. For usage details, see the package vignette ("BayesPIM_intro"). Further details can be found in T. Klausch, B. I. Lissenberg-Witte, and V. M. Coupe (2024), "A Bayesian prevalence-incidence mixture model for screening outcomes with misclassification", \doi{10.48550/arXiv.2412.16065}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasklausch2/bayespim}
  \item Report bugs at \url{https://github.com/thomasklausch2/BayesPIM/issues}
}

}
\author{
\strong{Maintainer}: Thomas Klausch \email{t.klausch@amsterdamumc.nl}

}
\keyword{internal}
