% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPIM_package.r
\docType{package}
\name{BayesPIM-package}
\alias{BayesPIM}
\alias{BayesPIM-package}
\title{BayesPIM: Bayesian Prevalence Incidence Mixture model}
\description{
Model time to event from interval censored (screening) data when there is latent prevalence at baseline and misclassification (imperfect test sensitivity).
}
\author{
\strong{Maintainer}: Thomas Klausch \email{t.klausch@amsterdamumc.nl}

}
\keyword{internal}
