% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPIM_package.r
\docType{package}
\name{BayesPIM-package}
\alias{BayesPIM}
\alias{BayesPIM-package}
\title{BayesPIM: Bayesian Prevalence-Incidence Mixture Model}
\description{
This package models time-to-event data from interval-censored screening studies. It accounts for latent prevalence at baseline and incorporates misclassification due to imperfect test sensitivity. For details on usage see the vignette("BayesPIM_intro").
}
\author{
\strong{Maintainer}: Thomas Klausch \email{t.klausch@amsterdamumc.nl}

}
\keyword{internal}
