% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datgen.r
\name{gen.dat}
\alias{gen.dat}
\title{gen.dat}
\usage{
gen.dat(
  kappa = 0.7,
  n = 1000,
  p = 2,
  p.discrete = 0,
  r = 0,
  s = 1,
  sigma.X = 1/2,
  mu.X = 4,
  beta.X = NULL,
  beta.W = NULL,
  theta = 0.15,
  Tmax = 20,
  v.min = 1,
  v.max = 6,
  mean.rc = 40,
  dist.X = "weibull",
  k = 1,
  sel.mod = "probit",
  prob.r = 0
)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
gen.dat
}
