% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.IC_2S.r
\name{get.IC_2S}
\alias{get.IC_2S}
\title{get.IC_2S}
\usage{
get.IC_2S(mod, samples = nrow(mod$par.X.bi[[1]]), cores = NULL)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
get.IC_2S
}
