% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ppd.2S.r
\name{get.ppd.2S}
\alias{get.ppd.2S}
\title{get.ppd.2S}
\usage{
get.ppd.2S(
  mod,
  fix_Z.X = NULL,
  fix_Z.W = NULL,
  pst.samples = 10^3,
  perc = seq(0, 1, 0.01),
  type = "x",
  ppd.type = "percentiles",
  quant = NULL
)
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
get.ppd.2S
}
