% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.prop.sd_seq.r
\name{search.prop.sd_seq}
\alias{search.prop.sd_seq}
\title{search.prop.sd_seq}
\usage{
search.prop.sd_seq(m, ndraws = 1000, succ.min = 3, acc.bounds.X = c(0.2, 0.25))
}
\arguments{
\item{x}{Numeric; Description of x}
}
\value{
Numeric; Description of the return value
}
\description{
search.prop.sd_seq
}
